+++ src/main/java/com/dabman/WebAPI/services/ProductService.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 public interface ProductService {
-    public Product getProductByid(int id);
+    public Product getProductByid(Long id);
     public List<Product> getAllProducts();
 }
